requirements.txt
streamlit
pandas
python-dateutil
openai==0.28.0
python-dotenv
